Reworked modal and changed fav icon on movies, the like button is now in the modal in searchbar
