+++ frontend/src/components/Category.scss
@@ -15,7 +15,7 @@
 figure.category {
   display: inline-block;
   width: 15rem;
-  height: 18rem;
+  // height: 18rem;
   margin: 0.5rem;
   margin-top: 5rem;
   flex-wrap: wrap;
