+++ frontend/src/components/CoupDeCoeur.css
@@ -24,7 +24,7 @@
 @media screen and (min-width: 1024px) {
   .cc {
     display: flex;
-    width: 30%;
+    width: 20%;
     flex-direction: column;
     border-radius: 30px;
     background-color: rgb(0, 255, 255);
