+++ frontend/src/components/DisplayComponent.css
@@ -5,10 +5,16 @@
   margin: 0;
 }
 
+.contenu {
+  font-size: 0.7rem;
+
+}
+
 @media screen and (max-width: 850px) {
   .contenu {
     display: flex;
     flex-direction: column;
+    
   }
 
   .date {
