+++ frontend/src/components/Modal.css
@@ -13,14 +13,27 @@
   flex-direction: row;
   align-items: center;
   justify-content: space-around;
+  padding-bottom: 20px;
+  
+}
+
+.ccmodal {
+  display: flex;
+  width: 10%;
+  font-size: 1.5rem;
+  flex-direction: column;
+  border-radius: 30px;
+  background-color: rgb(0, 255, 255);
+  margin: 0%;
+  align-items: center;
+  
 }
 
 .modalimg {
-  padding-left: 10px;
-  padding-top: 10px;
-  padding-bottom: 10px;
+  padding : 10px;
   border-radius: 4%;
-  max-height: 90%;
+  max-height: 70%;
+
 }
 
 .modal-body {
@@ -33,8 +46,8 @@
   padding-right: 10px;
   font-size: 1.3rem;
   display: flex;
-  flex-direction: column;
-  align-items: flex-end;
+  justify-content: space-between;
+  align-items: center;
 }
 .close {
   color: blue;
@@ -83,6 +96,7 @@
   background: linear-gradient(to top, #63b8ee, #468ccf);
   display: flex;
   justify-content: center;
+  align-self: flex-end;
 }
 .modal-close-button:hover {
   color: #14396a !important;
@@ -96,20 +110,21 @@
 .modal-overview {
   white-space: break-spaces;
   font-size: 1.3rem;
+  font-weight: 800;
   width: 100%;
   max-width: 55ch;
   color: rgb(255, 255, 255);
   padding-left: 10px;
 }
 
-@media screen and (max-width: 850px) {
+@media screen and (max-width: 950px) {
   .modal-body {
     max-width: fit-content;
     font-size: 3rem;
   }
   .modal-overlay {
     display: flex;
-    justify-content: space-between;
+    justify-content: space-evenly;
     flex-direction: column;
     max-height: calc(100vh - 0px);
     overflow-y: auto;
@@ -138,5 +153,6 @@
     background-size: contain;
     background-repeat: no-repeat;
     height: auto;
+    align-self: center;
   }
 }
