+++ frontend/src/components/Modal.jsx
@@ -1,16 +1,42 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./Modal.css";
+import { AiTwotoneHeart } from "react-icons/ai";
 
-const Modal = ({ isShowing, hide, title, overview, posterPath, vote, date }) =>
+const addStorage = (id) => {
+  const storedData = window.localStorage.id
+    ? window.localStorage.id.split(",")
+    : [];
+
+  if (!storedData.includes(id.toString())) {
+    storedData.push(id);
+    window.localStorage.id = storedData;
+  }
+};
+const Modal = ({
+  isShowing,
+  hide,
+  title,
+  overview,
+  posterPath,
+  vote,
+  date,
+  id,
+}) =>
   isShowing
     ? ReactDOM.createPortal(
         <div>
           <div className="modal-overlay">
             <img className="modalimg" src={posterPath} alt="a" />
             <div className="flexprops">
+              <button
+                className="modal-close-button"
+                type="button"
+                onClick={hide}
+              >
+                X
+              </button>
               <h2 className="modal-body">{title}</h2>
-
               <p className="modal-overview">{overview}</p>
               <div className="flexbot">
                 <span style={{ color: vote >= 7 ? "green" : "orange" }}>
@@ -18,11 +44,14 @@ const Modal = ({ isShowing, hide, title, overview, posterPath, vote, date }) =>
                 </span>
                 <p className="modal-date">Release date : {date}</p>
                 <button
-                  className="modal-close-button"
+                  className="ccmodal"
                   type="button"
-                  onClick={hide}
+                  onClick={() => {
+                    addStorage(id);
+                    window.location.reload();
+                  }}
                 >
-                  X
+                  <AiTwotoneHeart />
                 </button>
               </div>
             </div>
