+++ frontend/src/components/Populars.css
@@ -43,6 +43,7 @@ h2 {
 }
 .contenu {
   padding: 5px;
+  
 }
 
 /* coup de coeur */
